[port] codegen: Bun.build → esbuild in bundle-functions.ts
codegen/bundle-functions: don't keepNames

esbuild emits helper declarations at the top of global scope which are called within the capture blocks but not defined within the capture blocks, which breaks.

bun seems to always define the function as anonymous anyway, so we only lose named inners with this change.

a more proper fix would be to possibly restructure the capture blocks to contain top-level definitions (inside an IIFE), or to disable esbuild name minification in the first place